refactor(#792): decompose chain_row_wiring::wire (470-line fn) by responsibility
The single 470-line wire() becomes an orchestrator + 4 named helpers grouped by
responsibility: wire_delete_flow, wire_chain_mutations (base file) and wire_reorder,
wire_di_loop (chain_row_wiring_actions.rs sibling). Behaviour identical — same closures,
same captures via &ChainRowCtx. File 621->415, off debt. adapter-gui chain_row tests: 12 pass.
